Make the CinnamonGlobal struct available to C code
This makes it easier to reduce public function invocation.
